+++ Rave/Rave/Engine/Graphics/Include/ImageDecoder.h
@@ -12,7 +12,7 @@ namespace rave
 		void ReadImage(const wchar_t* filename, UINT& width, UINT& height, const REFWICPixelFormatGUID format);
 		void CopyPixels(void* pData, const UINT size, const UINT stride);
 
-		void SaveImage(std::wstring_view filename, const UINT width, const UINT height, const UINT size, const UINT stride, WICPixelFormatGUID format, void* pData, const size_t nPaletteColors = 256, const bool useAlpha = true);
+		void SaveImage(std::wstring_view filename, const UINT width, const UINT height, const UINT size, const UINT stride, WICPixelFormatGUID format, void* pData, const UINT nPaletteColors = 256, const bool useAlpha = true);
 
 	private:
 		ComPtr<IWICImagingFactory>	pFactory;
